Pruned navigation package
