+++ riptide_navigation/package.xml
@@ -26,6 +26,8 @@
   <run_depend>riptide_msgs</run_depend>
   <run_depend>sensor_msgs</run_depend>
   <run_depend>geometry_msgs</run_depend>
+
   <export>
   </export>
+  
 </package>
